IncrementalStore: Fixed 64-bit compile error. Moved out of ‘iOS’ group in project tree.
CBLIncrementalStore should work on Mac too.

+++ Source/API/Extras/CBLIncrementalStore.m
@@ -587,7 +587,7 @@ static NSString *CBLISResultTypeName(NSFetchRequestResultType resultType);
         if (outError) *outError = [NSError errorWithDomain:kCBLISErrorDomain
                                                       code:CBLIncrementalStoreErrorUnsupportedRequestType
                                                   userInfo:@{
-                                                             NSLocalizedFailureReasonErrorKey: [NSString stringWithFormat:@"Unsupported requestType: %d", request.requestType]
+                                                             NSLocalizedFailureReasonErrorKey: [NSString stringWithFormat:@"Unsupported requestType: %d", (int)request.requestType]
                                                              }];
         return nil;
     }
